+++ src/core/geometry/Arc3D.cpp
@@ -29,7 +29,7 @@ void Arc3D_Geo::buildVertexGeometries()
     osg::ref_ptr<osg::Vec3Array> vertices = new osg::Vec3Array;
 
     // 收集所有控制点
-    std::vector<glm::vec3> allPoints;
+    std::vector<glm::dvec3> allPoints;
     for (auto& points : controlPointss)
         for (auto& point : points)
         {
@@ -75,7 +75,7 @@ void Arc3D_Geo::buildEdgeGeometries()
     }
 
     // 收集所有控制点
-    std::vector<glm::vec3> allPoints;
+    std::vector<glm::dvec3> allPoints;
     for (auto& points : controlPointss)
         for (auto& point : points)
         {
@@ -140,4 +140,7 @@ void Arc3D_Geo::buildFaceGeometries()
     mm_node()->clearFaceGeometry();
     // 圆弧没有面
 }
+
+
+
